+++ embassy-mspm0/src/tim/input_capture.rs
@@ -421,8 +421,6 @@ impl<W: Word> CaptureChannel<'_, W> {
 
 impl<T: Instance> Drop for InputCapture<'_, T> {
     fn drop(&mut self) {
-        for pin in self.pins.iter().filter_map(MaybeAnyPin::pin) {
-            pin.set_as_disconnected();
-        }
+        crate::tim::disconnect_pins(&self.pins);
     }
 }